requirements.txt
keyboard==0.13.5
PyPDF2==3.0.1
pyperclip==1.8.2
python-dotenv==1.0.1